Rename file and merge ReHo and ALFF tests.
